requirements.txt
pandas
SQLAlchemy
psycopg2-binary
pytestpytest
